move norm, mask functions into utils
separate files for "frames" in coi loss corr

shorter var names

added default fixtures
