+++ dresscode/corrections.py
@@ -16,8 +16,11 @@ from astropy.io import fits
 from astropy.io.fits.hdu.hdulist import HDUList
 
 from dresscode.utils import (
+    apply_mask,
     check_filter,
     load_config,
+    norm,
+    update_mask,
     windowed_finite_vals,
     windowed_std,
     windowed_sum,
@@ -53,131 +56,54 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
         fname = path + fname
 
         # update the masks
+        # remove pixels that are NaN in the exposure map and pixels that have very low exposure times
         mask_fname = fname.replace("_sk_corr.img", "_mk_corr.img")
         exp_fname = mask_fname.replace("mk", "ex")
         new_mask_fname = mask_fname.replace(".img", "_new.img")
-        new_mask_data = update_mask(mask_fname, exp_fname, new_mask_fname)
+        new_mask_hdul = update_mask(mask_fname, exp_fname, new_mask_fname)
 
         # we manipulate the data directly, so open it in memory
-        hdul_unmasked = fits.open(fname)
+        unmasked_hdul = fits.open(fname)
 
         # apply mask to data, set 0's in mask to nan's
         # coincidence loss correction factor & uncertainties need to take into account missing data
         masked_hdul_fname = fname.replace(".img", "_mk.img")
-        hdul_masked = apply_mask(hdul_unmasked, new_mask_data, masked_hdul_fname)
+        masked_hdul = apply_mask(unmasked_hdul, new_mask_hdul, masked_hdul_fname)
 
         # apply normalization to data to convert to counts/sec
         norm_hdul_fname = fname.replace(".img", "_nm.img")
         exp_hdul = fits.open(exp_fname)
-        hdul_norm = norm(hdul_masked, exp_hdul, norm_hdul_fname)
+        norm_hdul = norm(masked_hdul, exp_hdul, norm_hdul_fname)
 
-        # run corrections on data, saving "planes" as separate files
-
-        # Apply a coincidence loss correction.
+        # Apply a coincidence loss correction, saving "planes" as separate files
         print("Applying coincidence loss corrections...")
-        coicorr_hdulist, coicorr_filename = coicorr(hdul_norm, fname)
+        coicorr_hdul, coicorr_fname = coicorr(norm_hdul, fname)
 
         # Apply a large scale sensitivity correction.
         print("Applying large scale sensitivity corrections...")
-        lsscorr_hdulist, lsscorr_filename = lsscorr(coicorr_hdulist, coicorr_filename)
+        lsscorr_hdul, lsscorr_fname = lsscorr(coicorr_hdul, coicorr_fname)
 
-        # Apply a zero point correction.
-        # todo: separate file for zero point correction
+        # Apply a zero point correction
         print("Applying zero point corrections...")
-        zp_corr_hdulist, _ = zeropoint(
-            lsscorr_hdulist, lsscorr_filename, *ZEROPOINT_PARAMS[check_filter(fname)]
+        zp_corr_hdul, zp_corr_fname = zeropoint(
+            lsscorr_hdul, lsscorr_fname, *ZEROPOINT_PARAMS[check_filter(fname)]
         )
 
-        # requirement of coadding step, data must be 2D
-        print(
-            "Splitting the corrected data, coincidence corr, and coicorr_rel into separate files"
-        )
-        separate_files(zp_corr_hdulist, fname)
+        # remove normalization and convert back to counts (needed for uvotimsum)
+        denorm_hdul_fname = zp_corr_fname.replace(".img", "_dn.img")
+        norm(zp_corr_hdul, exp_hdul, denorm_hdul_fname, denorm=True)
 
         print(f"Corrected image {i + 1}/{len(filenames)}.")
 
     return 0
 
 
-def apply_mask(
-    hdulist: HDUList, mask: HDUList, output_fname: str, dry_run: bool = False
-) -> HDUList:
-    """apply the mask to the image frames"""
-    new_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
-    new_hdulist = fits.HDUList([new_hdu_header])
-
-    for frame, mask_frame in zip(hdulist[1:], mask[1:]):
-        new_frame = np.full_like(frame.data, np.nan)
-        data_vals = mask_frame.data.astype(bool)
-        new_frame[data_vals] = frame.data[data_vals]
-        new_hdu = fits.ImageHDU(new_frame, frame.header)
-        new_hdulist.append(new_hdu)
-
-    if not dry_run:
-        # save the hdulist to a new file
-        new_hdulist.writeto(output_fname)
-
-    return new_hdulist
-
-
-def update_mask(
-    mask_fname: str, exp_fname: str, output_fname: str, dry_run: bool = False
-) -> HDUList:
-    """update the mask with pixels that are NaN in the exposure map and pixels
-    that have very low exposure times."""
-
-    # Open the mask file and the exposure map and copy the primary header (extension 0
-    # of hdulist) to a new hdulist
-    hdulist_mk = fits.open(mask_fname)
-    hdulist_ex = fits.open(exp_fname)
-    new_hdu_header = fits.PrimaryHDU(header=hdulist_mk[0].header)
-    new_hdulist = fits.HDUList([new_hdu_header])
-
-    for mk_frame, ex_frame in zip(hdulist_mk[1:], hdulist_ex[1:]):
-        # set mask pixels to 0 that are NaN in the exposure map or whose exposure time is very small
-        new_mask = mk_frame.data * np.isfinite(ex_frame.data) * (ex_frame.data > 1.0)
-        new_hdu = fits.ImageHDU(new_mask, mk_frame.header)
-        new_hdulist.append(new_hdu)
-
-    # Write the new hdulist to new mask file
-    if not dry_run:
-        new_hdulist.writeto(output_fname)
-        print(
-            f"{os.path.basename(output_fname)} has been updated with the exposure map"
-        )
-
-    return new_hdulist
-
-
-def norm(
-    data_hdul: HDUList, exp_hdul: HDUList, output_fname: str, dry_run: bool = False
-) -> HDUList:
-    """normalize the data by the exposure map"""
-    new_hdu_header = fits.PrimaryHDU(header=data_hdul[0].header)
-    new_hdulist = fits.HDUList([new_hdu_header])
-
-    for data_frame, exp_frame in zip(data_hdul[1:], exp_hdul[1:]):
-        new_frame = np.full_like(data_frame.data, np.nan)
-        finite_vals = np.isfinite(data_frame.data) * np.isfinite(exp_frame.data)
-        new_frame[finite_vals] = (
-            data_frame.data[finite_vals] / exp_frame.data[finite_vals]
-        )
-        new_hdu = fits.ImageHDU(new_frame, data_frame.header)
-        new_hdulist.append(new_hdu)
-
-    if not dry_run:
-        # save the hdulist to a new file
-        new_hdulist.writeto(output_fname)
-        print(f"{os.path.basename(output_fname)} normalized")
-
-    return new_hdulist
-
-
-# Functions for PART 1
-def coicorr(hdulist, filename):
+def coicorr(hdulist: HDUList, fname: str):
     """Coincidence loss correction"""
     new_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
-    new_hdulist = fits.HDUList([new_hdu_header])
+    coi_loss_corr_hdl = fits.HDUList([new_hdu_header])
+    corrfactor_hdl = fits.HDUList([new_hdu_header])
+    coicorr_unc_hdl = fits.HDUList([new_hdu_header])
 
     for frame in hdulist[1:]:
         data = frame.data
@@ -192,14 +118,13 @@ def coicorr(hdulist, filename):
         # standard deviation of the flux densities in the 9x9 pixels box.
         std = windowed_std(data, radius, win_finite_vals=window_pixels)
 
-        # Obtain the dead time correction factor and the frame time (in s) from the header
-        # of the image.
+        # Get dead time correction factor and frame time (in s)
         alpha = header["DEADC"]
-        ft = header["FRAMTIME"]
+        ft = header["FRAMTIME"]  # normally 11 ms, time between readouts
 
         # Calculate the total number of counts in the 9x9 pixels box: x = Craw*ft (counts).
-        # Calculate the minimum and maximum possible number of counts in the 9x9 pixels box.
         total_counts = ft * total_flux
+        # Calculate the minimum and maximum possible number of counts in the 9x9 pixels box.
         total_counts_min = ft * (total_flux - window_pixels * std)
         total_counts_max = ft * (total_flux + window_pixels * std)
 
@@ -254,7 +179,7 @@ def coicorr(hdulist, filename):
 
         print(
             "The median coincidence loss correction factor for image "
-            + os.path.basename(filename)
+            + os.path.basename(fname)
             + " is "
             + str(np.nanmedian(corrfactor))
             + " and the median relative uncertainty on the corrected data is "
@@ -262,23 +187,26 @@ def coicorr(hdulist, filename):
             + "."
         )
 
-        # Adapt the header. Write the corrected data, the applied coincidence loss
-        # correction and the relative uncertainty to a new image.
-        header["PLANE0"] = "primary (counts/s)"
-        header["PLANE1"] = "coincidence loss correction factor"
-        header["PLANE2"] = "relative coincidence loss correction uncertainty (fraction)"
+        # todo: should we keep these in the header somewhere?
+        # header["PLANE0"] = "primary (counts/s)"
+        # header["PLANE1"] = "coincidence loss correction factor"
+        # header["PLANE2"] = "relative coincidence loss correction uncertainty (fraction)"
 
-        datacube = [new_data, corrfactor, coicorr_rel]
+        # Write the corrected data, coincidence loss correction factor, and relative uncertainty
+        coi_loss_corr_hdl.append(fits.ImageHDU(new_data, header))
+        corrfactor_hdl.append(fits.ImageHDU(corrfactor, header))
+        coicorr_unc_hdl.append(fits.ImageHDU(coicorr_unc, header))
 
-        new_hdu = fits.ImageHDU(datacube, header)
-        new_hdulist.append(new_hdu)
+    new_fname = fname.replace(".img", "_coi.img")
+    coi_loss_corr_hdl.writeto(new_fname, overwrite=True)
+    corrfactor_hdl.writeto(new_fname.replace(".img", "_corrfactor.img"), overwrite=True)
+    coicorr_unc_hdl.writeto(
+        new_fname.replace(".img", "_coicorr_unc.img"), overwrite=True
+    )
 
-    new_filename = filename.replace(".img", "_coi.img")
-    new_hdulist.writeto(new_filename, overwrite=True)
+    print(f"{os.path.basename(fname)} has been corrected for coincidence loss.")
 
-    print(os.path.basename(filename) + " has been corrected for coincidence loss.")
-
-    return new_hdulist, new_filename
+    return coi_loss_corr_hdl, new_fname
 
 
 def polynomial(x):
@@ -294,57 +222,53 @@ def polynomial(x):
     return 1 + (a1 * x) + (a2 * x**2) + (a3 * x**3) + (a4 * x**4)
 
 
-# Function for PART 2: Large scale sensitivity correction.
-def lsscorr(hdulist, filename):
+def lsscorr(hdulist: HDUList, fname: str):
+    """Large scale sensitivity correction"""
 
     new_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
     new_hdulist = fits.HDUList([new_hdu_header])
 
-    lss_hdulist = fits.open(filename.replace("sk_corr_coi.img", "lss_corr.img"))
+    lss_hdulist = fits.open(fname.replace("sk_corr_coi.img", "lss_corr.img"))
 
     for frame, lss_frame in zip(hdulist[1:], lss_hdulist[1:]):
 
-        data = frame.data[0]
-        coicorr = frame.data[1]
-        coicorr_rel = frame.data[2]
+        data = frame.data
         header = frame.header
 
-        # Apply the large scale sensitivity correction to the data.
+        # Apply the large scale sensitivity correction to the data
         new_data = data / lss_frame.data
-        datacube = [new_data, coicorr, coicorr_rel]
 
-        new_hdu = fits.ImageHDU(datacube, header)
+        new_hdu = fits.ImageHDU(new_data, header)
         new_hdulist.append(new_hdu)
 
     # Write the corrected data to a new image.
-    new_filename = filename.replace(".img", "_lss.img")
-    new_hdulist.writeto(new_filename, overwrite=True)
+    new_fname = fname.replace(".img", "_lss.img")
+    new_hdulist.writeto(new_fname, overwrite=True)
 
     print(
-        os.path.basename(filename)
+        os.path.basename(fname)
         + " has been corrected for large scale sensitivity variations."
     )
 
-    return new_hdulist, new_filename
+    return new_hdulist, new_fname
 
 
-# Function for PART 3: Zero point correction.
-def zeropoint(hdulist, filename, param1, param2):
+def zeropoint(hdulist: HDUList, fname: str, param1: float, param2: float):
+    """Zero point correction (sensitivity loss over time)"""
 
     new_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
     new_hdulist = fits.HDUList([new_hdu_header])
 
     for frame in hdulist[1:]:
 
-        data = frame.data[0]
-        coicorr = frame.data[1]
-        coicorr_rel = frame.data[2]
+        data = frame.data
         header = frame.header
 
         obs_date = datetime.fromisoformat(header["DATE-OBS"]).date()
         # Calculate the number of years that have elapsed since the 1st of January 2005.
         first_date = date(2005, 1, 1)
         elapsed_time = obs_date - first_date
+        # todo: update to correctly account for leap years
         years_passed = elapsed_time.days / 365.25
 
         # Calculate the zero point correction.
@@ -356,65 +280,18 @@ def zeropoint(hdulist, filename, param1, param2):
         # Adapt the header.
         header["ZPCORR"] = zerocorr
 
-        # Write the corrected data to a new image.
-        datacube = [new_data, coicorr, coicorr_rel]
-        new_hdu = fits.ImageHDU(datacube, header)
+        # Write the corrected data to a new image
+        new_hdu = fits.ImageHDU(new_data, header)
         new_hdulist.append(new_hdu)
 
-    new_filename = filename.replace(".img", "_zp.img")
-    new_hdulist.writeto(new_filename, overwrite=True)
+    new_fname = fname.replace(".img", "_zp.img")
+    new_hdulist.writeto(new_fname, overwrite=True)
 
     print(
-        os.path.basename(filename)
-        + " has been corrected for sensitivity loss of the detector over time."
+        f"{os.path.basename(fname)} has been corrected for sensitivity loss of the detector over time."
     )
 
-    return new_hdulist, new_filename
-
-
-def separate_files(hdulist, filename):
-    """Separate data into different files for data, coicorr and coicorr_rel."""
-    data_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
-    data_hdulist = fits.HDUList([data_hdu_header])
-
-    coicorr_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
-    coicorr_hdulist = fits.HDUList([coicorr_hdu_header])
-
-    coicorr_rel_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
-    coicorr_rel_data_hdulist = fits.HDUList([coicorr_rel_hdu_header])
-    for frame in hdulist[1:]:
-
-        frame_header = frame.header
-        frame_header.remove("PLANE1")
-        frame_header.remove("PLANE2")
-
-        data_header = frame_header.copy()
-        data_header["PLANE0"] = "primary (counts/s)"
-        data = frame.data[0]
-        data_hdulist.append(fits.ImageHDU(data, data_header))
-
-        coicorr_header = frame_header.copy()
-        coicorr_header["PLANE0"] = "coincidence loss correction factor"
-        coicorr = frame.data[1]
-        coicorr_hdulist.append(fits.ImageHDU(coicorr, coicorr_header))
-
-        coicorr_rel_header = frame_header.copy()
-        coicorr_rel_header[
-            "PLANE0"
-        ] = "relative coincidence loss correction uncertainty (fraction)"
-        coicorr_rel = frame.data[2]
-        coicorr_rel_data_hdulist.append(fits.ImageHDU(coicorr_rel, coicorr_rel_header))
-
-    data_filename = filename.replace("coi_lss_zp.img", "coi_lss_zp_data.img")
-    data_hdulist.writeto(data_filename, overwrite=True)
-
-    coicorr_filename = filename.replace("coi_lss_zp.img", "coi_lss_zp_coicorr.img")
-    coicorr_hdulist.writeto(coicorr_filename, overwrite=True)
-
-    coicorr_rel_filename = filename.replace(
-        "coi_lss_zp.img", "coi_lss_zp_coicorr_rel.img"
-    )
-    coicorr_rel_data_hdulist.writeto(coicorr_rel_filename, overwrite=True)
+    return new_hdulist, new_fname
 
 
 if __name__ == "__main__":
