+++ dresscode/utils.py
@@ -4,6 +4,8 @@ import os
 
 import numpy as np
 from astropy.convolution import convolve
+from astropy.io import fits
+from astropy.io.fits.hdu.hdulist import HDUList
 
 
 def listdir_nohidden(path):
@@ -100,3 +102,81 @@ def windowed_finite_vals(arr: np.ndarray, radius: int) -> np.ndarray:
     output = windowed_sum(finite_arr, radius)
 
     return output
+
+
+def apply_mask(
+    hdulist: HDUList, mask: HDUList, output_fname: str, dry_run: bool = False
+) -> HDUList:
+    """apply the mask to the image frames"""
+    new_hdu_header = fits.PrimaryHDU(header=hdulist[0].header)
+    new_hdulist = fits.HDUList([new_hdu_header])
+
+    for frame, mask_frame in zip(hdulist[1:], mask[1:]):
+        new_frame = np.full_like(frame.data, np.nan)
+        data_vals = mask_frame.data.astype(bool)
+        new_frame[data_vals] = frame.data[data_vals]
+        new_hdu = fits.ImageHDU(new_frame, frame.header)
+        new_hdulist.append(new_hdu)
+
+    if not dry_run:
+        # save the hdulist to a new file
+        new_hdulist.writeto(output_fname)
+
+    return new_hdulist
+
+
+def update_mask(
+    mask_fname: str, exp_fname: str, output_fname: str, dry_run: bool = False
+) -> HDUList:
+    """update the mask with pixels that are NaN in the exposure map and pixels
+    that have very low exposure times."""
+
+    # Open the mask file and the exposure map and copy the primary header (extension 0
+    # of hdulist) to a new hdulist
+    hdulist_mk = fits.open(mask_fname)
+    hdulist_ex = fits.open(exp_fname)
+    new_hdu_header = fits.PrimaryHDU(header=hdulist_mk[0].header)
+    new_hdulist = fits.HDUList([new_hdu_header])
+
+    for mk_frame, ex_frame in zip(hdulist_mk[1:], hdulist_ex[1:]):
+        # set mask pixels to 0 that are NaN in the exposure map or whose exposure time is very small
+        new_mask = mk_frame.data * np.isfinite(ex_frame.data) * (ex_frame.data > 1.0)
+        new_hdu = fits.ImageHDU(new_mask, mk_frame.header)
+        new_hdulist.append(new_hdu)
+
+    # Write the new hdulist to new mask file
+    if not dry_run:
+        new_hdulist.writeto(output_fname)
+        print(f"Updated exposure map in {os.path.basename(output_fname)}")
+
+    return new_hdulist
+
+
+def norm(
+    data_hdul: HDUList,
+    exp_hdul: HDUList,
+    output_fname: str,
+    denorm: bool = False,
+    dry_run: bool = False,
+) -> HDUList:
+    """normalize the data by the exposure map"""
+    new_hdu_header = fits.PrimaryHDU(header=data_hdul[0].header)
+    new_hdulist = fits.HDUList([new_hdu_header])
+
+    for data_frame, exp_frame in zip(data_hdul[1:], exp_hdul[1:]):
+        new_frame = np.full_like(data_frame.data, np.nan)
+        finite_vals = np.isfinite(data_frame.data) * np.isfinite(exp_frame.data)
+        if denorm:
+            norm_factor = 1.0 / exp_frame.data[finite_vals]
+        else:
+            norm_factor = exp_frame.data[finite_vals]
+        new_frame[finite_vals] = data_frame.data[finite_vals] / norm_factor
+        new_hdu = fits.ImageHDU(new_frame, data_frame.header)
+        new_hdulist.append(new_hdu)
+
+    if not dry_run:
+        # save the hdulist to a new file
+        new_hdulist.writeto(output_fname)
+        print(f"{os.path.basename(output_fname)} normalized")
+
+    return new_hdulist
