+++ tests/test_utils.py
@@ -1,4 +1,11 @@
+from __future__ import annotations
+
+import os
+import subprocess
+
 import numpy as np
+import pytest
+from astropy.io import fits
 
 from dresscode import utils
 
@@ -218,3 +225,110 @@ def test_windowed_finite_vals_larger_arr():
     output_loop = windowed_finite_vals_loop(data, 1)
     test_output = utils.windowed_finite_vals(data, 1)
     assert np.array_equal(output_loop, test_output)
+
+
+@pytest.fixture
+def updated_mask_fname(mask_fname, exp_fname):
+    new_mask_fname = mask_fname.replace(".img", "_new.img")
+    utils.update_mask(mask_fname, exp_fname, new_mask_fname)
+    return new_mask_fname
+
+
+@pytest.fixture
+def data_masked_fname(data_fname: str, updated_mask_fname: str) -> str:
+    hdulist = fits.open(data_fname)
+    mask_hdulist = fits.open(updated_mask_fname)
+    output_fname = data_fname.replace(".img", "_mk.img")
+    utils.apply_mask(hdulist, mask_hdulist, output_fname)
+    return output_fname
+
+
+def test_update_mask(mask_fname, exp_fname):
+    new_mask_fname = mask_fname.replace(".img", "_new.img")
+    new_mask = utils.update_mask(mask_fname, exp_fname, new_mask_fname, dry_run=True)
+
+    mask_data = new_mask[1].data
+    assert mask_data[0, 0] == 0
+    assert mask_data[1, 1] == 0
+    assert mask_data[2, 2] == 1
+
+
+def test_apply_mask(data_fname, mask_fname):
+    data_hdulist = fits.open(data_fname)
+    mask_hdulist = fits.open(mask_fname)
+
+    data_hdulist_masked = utils.apply_mask(data_hdulist, mask_hdulist, "", dry_run=True)
+
+    data = data_hdulist_masked[1].data
+    assert np.isnan(data[0, 0])
+    assert data[1, 1] == data_hdulist[1].data[1, 1]
+
+
+def norm_farith(fname: str, exp_fname: str):
+    """normalize an image by its exposure map
+
+    used only for testing, for comparison with norm function
+    """
+    path = os.path.dirname(fname) + "/"
+
+    # Specify the input files and the output file.
+    infil1 = fname + "+1"
+    infil2 = exp_fname + "+1"
+    outfil = fname.replace(".img", "_nm.img")
+
+    # Run farith with the specified parameters:
+    subprocess.call(
+        f"farith infil1={infil1} infil2={infil2} outfil={outfil} ops=div null=y",
+        cwd=path,
+        shell=True,
+    )
+
+    print(f"{os.path.basename(outfil)} normalized")
+    return outfil
+
+
+def test_norm(data_masked_fname: str, exp_fname: str):
+
+    data_hdul = fits.open(data_masked_fname)
+    exp_hdul = fits.open(exp_fname)
+    outfname = data_masked_fname.replace(".img", "_nm.img")
+    hdul_norm = utils.norm(data_hdul, exp_hdul, outfname, dry_run=True)
+
+    data = hdul_norm[1].data
+    assert np.isnan(data[0, 0])  # masked
+    assert np.isnan(data[1, 1])  # zero exposure, should be masked in the data
+    assert data[1, 2] == data_hdul[1].data[1, 2] / 8
+
+
+@pytest.mark.skipif(
+    os.getenv("CI") == "true", reason="skip on CI, where we don't have farith installed"
+)
+def test_norm_farith(data_masked_fname: str, exp_fname: str):
+    # todo: we could run these tests inside our docker image in CI
+    data_hdul = fits.open(data_masked_fname)
+    exp_hdul = fits.open(exp_fname)
+    outfname = data_masked_fname.replace(".img", "_nm.img")
+    hdul_norm = utils.norm(data_hdul, exp_hdul, outfname, dry_run=True)
+    data = hdul_norm[1].data
+
+    # compare against norm using farith function
+    outfname_farith = norm_farith(data_masked_fname, exp_fname)
+
+    data_farith = fits.open(outfname_farith)[0].data
+    assert np.array_equal(data, data_farith, equal_nan=True)
+
+
+def test_denorm(data_masked_fname: str, exp_fname: str):
+    data_hdul = fits.open(data_masked_fname)
+    exp_hdul = fits.open(exp_fname)
+    norm_fname = data_masked_fname.replace(".img", "_nm.img")
+    norm_hdul = utils.norm(data_hdul, exp_hdul, norm_fname, dry_run=True)
+
+    denorm_hdul_fname = norm_fname.replace(".img", "_dn.img")
+    denorm_hdul = utils.norm(
+        norm_hdul, exp_hdul, denorm_hdul_fname, denorm=True, dry_run=True
+    )
+    data_denorm = denorm_hdul[1].data
+
+    assert np.array_equal(data_denorm, data_hdul[1].data, equal_nan=True)
+    assert not np.array_equal(data_denorm, norm_hdul[1].data, equal_nan=True)
